[BBC-9055] Migrate Menu to MDX
